@@ -109,7 +121,7 @@ setup(
     name='cumin',
     packages=find_packages(exclude=['*.tests', '*.tests.*']),
     platforms=['GNU/Linux', 'BSD', 'MacOSX'],
-    python_requires='>=3.7',
+    python_requires='>=3.9',
     setup_requires=setup_requires,
     url='https://github.com/wikimedia/cumin',
     use_scm_version=True,
