@@ -3,17 +3,18 @@ import logging
 import os
 import subprocess
 
+from importlib.metadata import PackageNotFoundError, version
+
 import yaml
 
 from ClusterShell.NodeSet import NodeSet, RESOLVER_NOGROUP
-from pkg_resources import DistributionNotFound, get_distribution
 
 
 KERBEROS_KLIST = '/usr/bin/klist'
-try:  # TODO: use importlib.metadata once Python 3.7 support is dropped
-    __version__ = get_distribution(__name__).version
+try:
+    __version__ = version(__name__)
     """:py:class:`str`: the version of the current Cumin module."""
-except DistributionNotFound:  # pragma: no cover - this happens only if the package is not installed
+except PackageNotFoundError:  # pragma: no cover - this happens only if the package is not installed
     # Support the use case of the Debian building system where tests are run without installation
     if 'SETUPTOOLS_SCM_PRETEND_VERSION' in os.environ:
         __version__ = os.environ['SETUPTOOLS_SCM_PRETEND_VERSION']
