@@ -37,20 +37,6 @@ def test_hosts_selection():
     assert hosts in parsed[0].asDict()['hosts']
 
 
-class TestPuppetDBQueryV3:
-    """PuppetDB backend query test class for API version 3."""
-
-    def setup_method(self, _):
-        """Set an instance of PuppetDBQuery for each test."""
-        config = {'puppetdb': {'api_version': 3}}
-        self.query = puppetdb.PuppetDBQuery(config)  # pylint: disable=attribute-defined-outside-init
-
-    def test_instantiation(self):
-        """An instance of PuppetDBQuery should be an instance of BaseQuery."""
-        assert isinstance(self.query, BaseQuery)
-        assert self.query.url == 'https://localhost:443/v3/'
-
-
 class TestPuppetDBQueryV4:
     """PuppetDB backend query test class for API version 4."""
 
@@ -95,115 +81,6 @@ class TestPuppetDBQueryV4:
             self.query.endpoint = 'resources'
 
 
-def test_puppetdb_query_init_invalid():
-    """Instantiating PuppetDBQuery with an unsupported API version should raise InvalidQueryError."""
-    with pytest.raises(InvalidQueryError, match='Unsupported PuppetDB API version'):
-        puppetdb.PuppetDBQuery({'puppetdb': {'api_version': 99}})
-
-
-@mock.patch.object(puppetdb.PuppetDBQuery, '_api_call')
-class TestPuppetDBQueryBuildV3:
-    """PuppetDB backend API v3 query build test class."""
-
-    def setup_method(self, _):
-        """Set an instace of PuppetDBQuery for each test."""
-        config = {'puppetdb': {'api_version': 3}}
-        self.query = puppetdb.PuppetDBQuery(config)  # pylint: disable=attribute-defined-outside-init
-
-    def test_add_category_resource_parameter_regex(self, mocked_api_call):
-        """A resource's parameter query with a regex should raise InvalidQueryError."""
-        with pytest.raises(InvalidQueryError, match='Regex operations are not supported in PuppetDB'):
-            self.query.execute('R:resource%param ~ value.*')
-            assert not mocked_api_call.called
-
-    @pytest.mark.parametrize('query, expected', (
-        (  # No hosts
-            'host1!host1',
-            ''),
-        (  # Single host
-            'host',
-            '["or", ["=", "name", "host"]]'),
-        (  # Multiple hosts
-            'host[1-2]',
-            '["or", ["=", "name", "host1"], ["=", "name", "host2"]]'),
-        (  # Negated query
-            'not host[1-2]',
-            '["not", ["or", ["=", "name", "host1"], ["=", "name", "host2"]]]'),
-        (  # Globbing hosts
-            'host1*.domain',
-            r'["or", ["~", "name", "^host1.*\\.domain$"]]'),
-    ))
-    def test_add_hosts(self, mocked_api_call, query, expected):
-        """A host query should add the proper query token to the current_group."""
-        self.query.execute(query)
-        mocked_api_call.assert_called_with(expected)
-
-    @pytest.mark.parametrize('query, expected', (
-        (  # number
-            'R:resource = 5551723',
-            '["and", ["=", "type", "Resource"], ["=", "title", 5551723]]'),
-        (  # number (neg)
-            'R:resource = -23',
-            '["and", ["=", "type", "Resource"], ["=", "title", -23]]'),
-        (  # float
-            'R:resource = 8675.309',
-            '["and", ["=", "type", "Resource"], ["=", "title", 8675.309]]'),
-        (  # hex
-            'R:resource = 0x235A',
-            '["and", ["=", "type", "Resource"], ["=", "title", 9050]]'),
-        (  # lower hex
-            'R:resource = 0x235a',
-            '["and", ["=", "type", "Resource"], ["=", "title", 9050]]'),
-        (  # oct
-            'R:resource = 0777',
-            '["and", ["=", "type", "Resource"], ["=", "title", 511]]'),
-        (  # zero
-            'R:resource = 0',
-            '["and", ["=", "type", "Resource"], ["=", "title", 0]]'),
-        (  # bareword
-            'R:resource = true',
-            '["and", ["=", "type", "Resource"], ["=", "title", true]]'),
-        (  # quoted
-            'R:resource = "value words"',
-            '["and", ["=", "type", "Resource"], ["=", "title", "value words"]]'),
-        (  # unquoted
-            'R:resource = value',
-            '["and", ["=", "type", "Resource"], ["=", "title", "value"]]'),
-        (  # shortcut with numeric
-            'O:role_name%param > 0x235A',
-            ('["and", ["and", ["=", "type", "Class"], ["=", "title", "Role::Role_name"]], ["and", ["=", "type", "Class"'
-             '], [">", ["parameter", "param"], 9050]]]')),
-        (  # shortcut with bareword
-            'O:role_name%param = true',
-            ('["and", ["and", ["=", "type", "Class"], ["=", "title", "Role::Role_name"]], ["and", ["=", "type", "Class"'
-             '], ["=", ["parameter", "param"], true]]]')),
-    ))
-    def test_types(self, mocked_api_call, query, expected):
-        """A query should decode various types that puppetdb supports."""
-        self.query.execute(query)
-        mocked_api_call.assert_called_with(expected)
-
-    @pytest.mark.parametrize('query, operator, expected', (
-        (  # AND
-            'host1 and host2',
-            'and',
-            '["and", ["or", ["=", "name", "host1"]], ["or", ["=", "name", "host2"]]]'),
-        (  # OR
-            'host1 or host2',
-            'or',
-            '["or", ["or", ["=", "name", "host1"]], ["or", ["=", "name", "host2"]]]'),
-        (  # Multiple AND
-            'host1 and host2 and host3',
-            'and',
-            '["and", ["or", ["=", "name", "host1"]], ["or", ["=", "name", "host2"]], ["or", ["=", "name", "host3"]]]'),
-    ))
-    def test_operator(self, mocked_api_call, query, operator, expected):
-        """A query with boolean operators should set the boolean property to the current group."""
-        self.query.execute(query)
-        assert self.query.current_group['bool'] == operator
-        mocked_api_call.assert_called_with(expected)
-
-
 @mock.patch.object(puppetdb.PuppetDBQuery, '_api_call')
 class TestPuppetDBQueryBuildV4:
     """PuppetDB backend API v4 query build test class."""
@@ -231,6 +108,7 @@ class TestPuppetDBQueryBuildV4:
     ))
     def test_add_category_fact(self, mocked_api_call, query, expected):
         """A fact query should add the proper query token to the current_group."""
+        expected = f'["extract", ["certname"], {expected}, ["group_by", "certname"]]'
         self.query.execute(query)
         mocked_api_call.assert_called_with(expected)
 
@@ -313,6 +191,7 @@ class TestPuppetDBQueryBuildV4:
     ))
     def test_add_category_resource(self, mocked_api_call, query, expected):
         """A resource query should add the proper query token to the current_group."""
+        expected = f'["extract", ["certname"], {expected}, ["group_by", "certname"]]'
         self.query.execute(query)
         mocked_api_call.assert_called_with(expected)
 
@@ -376,6 +255,7 @@ class TestPuppetDBQueryBuildV4:
     ))
     def test_add_hosts(self, mocked_api_call, query, expected):
         """A host query should add the proper query token to the current_group."""
+        expected = f'["extract", ["certname"], {expected}, ["group_by", "certname"]]'
         self.query.execute(query)
         mocked_api_call.assert_called_with(expected)
 
@@ -397,6 +277,7 @@ class TestPuppetDBQueryBuildV4:
     ))
     def test_operator(self, mocked_api_call, query, operator, expected):
         """A query with boolean operators should set the boolean property to the current group."""
+        expected = f'["extract", ["certname"], {expected}, ["group_by", "certname"]]'
         self.query.execute(query)
         assert self.query.current_group['bool'] == operator
         mocked_api_call.assert_called_with(expected)
@@ -433,9 +314,8 @@ def test_complex_query(query_requests):
     """Calling execute() with a complex query should return the exptected structure."""
     category = 'R'
     endpoint = query_requests[0].endpoints[category]
-    key = query_requests[0].hosts_keys[endpoint]
-    query_requests[1].register_uri('GET', query_requests[0].url + endpoint + '?query=', status_code=200, json=[
-        {key: endpoint + '_host1', 'key': 'value1'}, {key: endpoint + '_host2', 'key': 'value2'}])
+    query_requests[1].register_uri('POST', query_requests[0].url + endpoint + '?query=', status_code=200, json=[
+        {'certname': endpoint + '_host1', 'key': 'value1'}, {'certname': endpoint + '_host2', 'key': 'value2'}])
 
     hosts = query_requests[0].execute('(resources_host1 or resources_host2) and R:Class = MyClass')
     assert hosts == nodeset('resources_host[1-2]')
